Functions

Functions

Functions exported from GeophysicalFlows:

Functions exported from TwoDTurb:

GeophysicalFlows.TwoDTurb.Problem — Method.

Problem(; parameters...)

Construct a 2D turbulence problem.

source

GeophysicalFlows.TwoDTurb.dissipation — Method.

dissipation(prob)
dissipation(s, v, p, g)

Returns the domain-averaged dissipation rate. nnu must be >= 1.

source

GeophysicalFlows.TwoDTurb.drag — Method.

drag(prob)
drag(s, v, p, g)

Returns the extraction of domain-averaged energy by drag/hypodrag mu.

source

GeophysicalFlows.TwoDTurb.energy — Method.

energy(prob)
energy(s, v, g)

Returns the domain-averaged kinetic energy in the Fourier-transformed vorticity solution s.sol.

source

GeophysicalFlows.TwoDTurb.enstrophy — Method.

enstrophy(s, g)

Returns the domain-averaged enstrophy in the Fourier-transformed vorticity solution s.sol.

source

GeophysicalFlows.TwoDTurb.set_q! — Method.

set_q!(prob, q)
set_q!(s, v, g, q)

Set the solution s.sol as the transform of q and update variables v on the grid g.

source

GeophysicalFlows.TwoDTurb.updatevars! — Method.

updatevars!(v, s, g)

Update the vars in v on the grid g with the solution in s.sol.

source

GeophysicalFlows.TwoDTurb.work — Method.

work(prob)
work(s, v, p, g)

Returns the domain-averaged rate of work of energy by the forcing Fh.

source

Functions exported from BarotropicQG:

GeophysicalFlows.BarotropicQG.Problem — Method.

Problem(; parameters...)

Construct a BarotropicQG turbulence problem.

source

GeophysicalFlows.BarotropicQG.dissipation — Method.

dissipation(prob)
dissipation(s, v, p, g)

Returns the domain-averaged dissipation rate. nnu must be >= 1.

source

GeophysicalFlows.BarotropicQG.drag — Method.

drag(prob)
drag(s, v, p, g)

Returns the extraction of domain-averaged energy by drag mu.

source

GeophysicalFlows.BarotropicQG.energy — Method.

Calculate the domain-averaged kinetic energy.

source

GeophysicalFlows.BarotropicQG.enstrophy — Method.

Returns the domain-averaged enstrophy.

source

GeophysicalFlows.BarotropicQG.meanenergy — Method.

Returns the energy of the domain-averaged U.

source

GeophysicalFlows.BarotropicQG.meanenstrophy — Method.

Returns the enstrophy of the domain-averaged U.

source

GeophysicalFlows.BarotropicQG.set_zeta! — Method.

set_zeta!(prob, zeta)
set_zeta!(s, v, g, zeta)

Set the solution s.sol as the transform of zeta and update variables v on the grid g.

source

GeophysicalFlows.BarotropicQG.updatevars! — Method.

updatevars!(v, s, g)

Update the vars in v on the grid g with the solution in s.sol.

source

GeophysicalFlows.BarotropicQG.work — Method.

work(prob)
work(s, v, p, g)

Returns the domain-averaged rate of work of energy by the forcing Fqh.

source

Functions exported from VerticallyFourierBoussinesq:

Functions exported from VerticallyCosineBoussinesq:

GeophysicalFlows.VerticallyCosineBoussinesq.mode0apv — Method.

mode0apv(prob)

Returns the barotropic available potential vorticity.

source

GeophysicalFlows.VerticallyCosineBoussinesq.mode0apv — Method.

mode0apv(uh, vh, ph, Zh, m, N, g)

Returns the barotropic available potential vorticity.

source

GeophysicalFlows.VerticallyCosineBoussinesq.mode0dissipation — Method.

mode0dissipation(prob)

Returns the domain-averaged barotropic dissipation rate. nnu0 must be >= 1.

source

GeophysicalFlows.VerticallyCosineBoussinesq.mode0drag — Method.

mode0drag(prob)

Returns the extraction of domain-averaged barotropic energy by drag μ.

source

GeophysicalFlows.VerticallyCosineBoussinesq.mode0energy — Method.

mode0energy(prob)

Returns the domain-averaged energy in the zeroth mode.

source

GeophysicalFlows.VerticallyCosineBoussinesq.mode0enstrophy — Method.

mode0enstrophy(prob)

Returns the domain-averaged enstrophy in the Fourier-transformed vorticity solution s.sol.

source

GeophysicalFlows.VerticallyCosineBoussinesq.mode1dissipation — Method.

mode1dissipation(prob)

Returns the domain-averaged kinetic energy dissipation of the first mode by horizontal viscosity.

source

GeophysicalFlows.VerticallyCosineBoussinesq.mode1drag — Method.

mode1drag(prob)

Returns the domain-averaged kinetic energy dissipation of the first mode by horizontal viscosity.

source

GeophysicalFlows.VerticallyCosineBoussinesq.mode1energy — Method.

mode1energy(prob)

Returns the domain-averaged total energy in the first mode.

source

GeophysicalFlows.VerticallyCosineBoussinesq.mode1ke — Method.

mode1ke(prob)

Returns the domain-averaged kinetic energy in the first mode.

source

GeophysicalFlows.VerticallyCosineBoussinesq.mode1pe — Method.

mode1pe(prob)

Returns the domain-averaged potential energy in the first mode.

source

GeophysicalFlows.VerticallyCosineBoussinesq.set_C! — Method.

set_C!(prob, C)

Set zeroth mode tracer concentration and update vars.

source

GeophysicalFlows.VerticallyCosineBoussinesq.set_Z! — Method.

set_Z!(prob, Z)

Set zeroth mode vorticity and update vars.

source

GeophysicalFlows.VerticallyCosineBoussinesq.set_planewave! — Function.

set_planewave!(prob, u₀, κ, θ=0)

Set a plane wave solution with initial speed u₀, non-dimensional wave number κ, and angle θ with the horizontal. The non-dimensional wavenumber vector is (k, l) = (κ cos θ, κ sin θ), is normalized by 2π/Lx, and is rounded to the nearest integer.

source

GeophysicalFlows.VerticallyCosineBoussinesq.set_uvp! — Method.

set_uvp!(prob)

Set first mode u, v, and p and update vars.

source

GeophysicalFlows.VerticallyCosineBoussinesq.totalenergy — Method.

totalenergy(prob)

Returns the total energy projected onto the zeroth mode.

source

GeophysicalFlows.VerticallyCosineBoussinesq.updatevars! — Method.

updatevars!(prob)

Update variables to correspond to the solution in s.sol or prob.state.sol.

source