Private types

Private types

Private types in module GeophysicalFlows:

Private types in module TwoDTurb:

GeophysicalFlows.TwoDTurb.ForcedVars — Method.

ForcedVars(g)

Returns the vars for forced two-dimensional turbulence with grid g.

source

GeophysicalFlows.TwoDTurb.Params — Type.

Params(nu, nnu, mu, nmu, calcF!)

Returns the params for two-dimensional turbulence.

source

GeophysicalFlows.TwoDTurb.StochasticForcedVars — Method.

StochasticForcedVars(g; T)

Returns the vars for stochastically forced two-dimensional turbulence with grid g.

source

GeophysicalFlows.TwoDTurb.Vars — Method.

Vars(g)

Returns the vars for unforced two-dimensional turbulence with grid g.

source

Private types in module BarotropicQG:

GeophysicalFlows.BarotropicQG.ForcedVars — Method.

ForcedVars(g)

Returns the vars for forced two-dimensional barotropic QG problem with grid g.

source

GeophysicalFlows.BarotropicQG.Params — Type.

Params(g::TwoDGrid, f0, beta, FU, eta, mu, nu, nnu, calcFU, calcFq)

Returns the params for an unforced two-dimensional barotropic QG problem.

source

GeophysicalFlows.BarotropicQG.Params — Method.

Params(g::TwoDGrid, f0, beta, eta::Function, mu, nu, nnu, calcFU, calcFq)

Constructor for Params that accepts a generating function for the topographic PV.

source

GeophysicalFlows.BarotropicQG.StochasticForcedVars — Method.

StochasticForcedVars(g)

Returns the vars for stochastically forced two-dimensional barotropic QG problem with grid g.

source

GeophysicalFlows.BarotropicQG.Vars — Method.

Vars(g)

Returns the vars for unforced two-dimensional barotropic QG problem with grid g.

source

Private types in module VerticallyFourierBoussinesq:

GeophysicalFlows.VerticallyFourierBoussinesq.Params — Type.

Params(nu0, nnu0, nu1, nnu1, f, N, m, Ub, Vb)

Construct parameters for the Two-Fourier-mode Boussinesq problem. Suffix 0 refers to zeroth mode; 1 to first mode. f, N, m are Coriolis frequency, buoyancy frequency, and vertical wavenumber of the first mode, respectively. The optional constant background velocity (Ub,Vb) is set to zero by default. The viscosity is applied only to the first-mode horizontal velocities.

source

Private types in module VerticallyCosineBoussinesq:

GeophysicalFlows.VerticallyCosineBoussinesq.ForcedVars — Method.

ForcedVars(g)

Returns the vars for forced two-vertical-cosine-mode Boussinesq dynamics on the grid g.

source

GeophysicalFlows.VerticallyCosineBoussinesq.Params — Type.

Params(nu0, nnu0, nu1, nnu1, mu0, nmu0, mu1, nmu1, f, N, m; Ub=0, Vb=0)

Construct parameters for the Two-Fourier-mode Boussinesq problem. Suffix 0 refers to zeroth mode; 1 to first mode. f, N, m are Coriolis frequency, buoyancy frequency, and vertical wavenumber of the first mode, respectively. The optional constant background velocity (Ub, Vb) is set to zero by default. The viscosity is applied only to the first-mode horizontal velocities.

source

GeophysicalFlows.VerticallyCosineBoussinesq.TracerForcedVars — Method.

TracerForcedVars(g)

Returns the vars for forced two-vertical-cosine-mode Boussinesq dynamics on the grid g.

source

GeophysicalFlows.VerticallyCosineBoussinesq.Vars — Method.

Vars(g)

Returns the vars for unforced two-vertical-cosine-mode Boussinesq dynamics on the grid g.

source