Private types

Private types in module GeophysicalFlows:

Private types in module TwoDTurb:

GeophysicalFlows.TwoDTurb.ForcedVars — Method

ForcedVars(dev, g)

Returns the vars for forced two-dimensional turbulence on device dev and with grid g.

GeophysicalFlows.TwoDTurb.Params — Type

Params(ν, nν, μ, nμ, calcF!)

Returns the params for two-dimensional turbulence.

GeophysicalFlows.TwoDTurb.StochasticForcedVars — Method

StochasticForcedVars(dev, g)

Returns the vars for stochastically forced two-dimensional turbulence on device dev and with grid g.

GeophysicalFlows.TwoDTurb.Vars — Method

Vars(dev, g)

Returns the vars for unforced two-dimensional turbulence on device dev and with grid g.

Private types in module BarotropicQG:

GeophysicalFlows.BarotropicQG.ForcedVars — Method

ForcedVars(dev, g)

Returns the vars for forced two-dimensional barotropic QG problem on device dev and with grid g.

GeophysicalFlows.BarotropicQG.Params — Type

Params(g::TwoDGrid, f0, β, FU, eta, μ, ν, nν, calcFU, calcFq)

Returns the params for an unforced two-dimensional barotropic QG problem.

GeophysicalFlows.BarotropicQG.Params — Method

Params(g::TwoDGrid, f0, β, eta::Function, μ, ν, nν, calcFU, calcFq)

Constructor for Params that accepts a generating function for the topographic PV.

GeophysicalFlows.BarotropicQG.StochasticForcedVars — Method

StochasticForcedVars(dev, g)

Returns the vars for stochastically forced two-dimensional barotropic QG problem on device dev and with grid g.

GeophysicalFlows.BarotropicQG.Vars — Method

Vars(dev, g)

Returns the vars for unforced two-dimensional barotropic QG problem on device dev and with grid g.